Refactor module and tests
