+++ lib/ama_layout.rb
@@ -9,6 +9,7 @@ require 'ama_layout/breadcrumb_builder'
 require 'ama_layout/moneris'
 require 'ama_layout/navigation'
 require 'ama_layout/navigation_item'
+require 'ama_layout/navigation_helper'
 require 'ama_layout/decorators/moneris_decorator'
 require 'ama_layout/decorators/navigation_decorator'
 require 'ama_layout/decorators/navigation_item_decorator'
